requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
pypdf
faiss-cpu
sentence-transformers
python-dotenv
